Thoughtfully use styles vs classes
In general, I'm liking the idea of using Bulma classes for things that
have more to do with layout (margins, flex) and CSS (via
styled-components) for things that are more purely "stylistic" like font
and color.

One edge case is that I'm using 'is-borderless' on the fieldset in
Slider. Border is technically more style than layout, but somehow it
feels like an okay exception to the rule. Since it's the only thing I
would have used styled-components for otherwise, it does save me an
import and simplify the component.

Another exception is that I'm using is-title for the h1 tag in the
header. I'm not sure I like the fact that Bulma resets all the h* tags
and then uses "title is-n" classes to make up for it. What I mean is
`<h1 class="title is-1">` feels redundant. I think the reason they do it
is so that they can use the same h1 tag as both a 'title' and a
'subtitle'. Whatever the reason, I'm just going with the flow since I'm
still learning.