Make fieldsets globally borderless
This somewhat resolves my conundrum in the previous commit about whether
to use is-borderless or styled-components to remove the border in the
Slider component. I just don't like the default border at all. Even
though this is a simple app and I'm not using fieldsets anywhere else
(and probably never will) it *feels* like a global preference. If I ever
did add another fieldset that wasn't a slider, I wouldn't want it to
have a border by default.