Fix H1 style in header
Now it feels a little extreme not to just use has-color-white but I
still like the consistency of only using utility classes for layout.

In reality, though, with scoped css the style is so close to the markup
that it's probably not *that* important to keep them separate.

"Keep components small and encapsulated" is probably more important than
"keep style and markup separate", even if both are good rules.